# Source/xrserver/static/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"description": "xr connect and aviation documentaion",
		"version": "1.0.0",
		"title": "XR Connect API"
	},
...
	"servers": [
		{
			"url": "/"
		}
	],
	"components": {
		"schemas": {
...
			  "addCompany": {
				"type": "object",
				"properties": {
					"companyid": {
						"type": "string"
					},
					"company_name": {
						"type": "string"
					},
                    "ceo_name": {
						"type": "string"

					},
                    "email":{
                        "type": "string"

                    },
                    "number":{
                        "type": "string"

                    },
                    "website": {
						"type": "string"
					},
                    "address": {
						"type": "string"
					},
					 "status": {
						"type": "string"
					},
					"created_by": {
						"type": "string"
					},
					"no_of_license": {
						"type": "string"
					},
					"license_key": {
						"type": "string"
					}
				}
			},
...
			"updateCompany": {
				"type": "object",
				"properties": {
					"companyid": {
						"type": "string"
					},
					"company_name": {
						"type": "string"
					},
                    "ceo_name": {
						"type": "string"

					},
                    "email":{
                        "type": "string"

                    },
                    "number":{
                        "type": "string"

                    },
                    "website": {
						"type": "string"
					},
                    "address": {
						"type": "string"
					},
					 "status": {
						"type": "string"
					}

				}
			},
			"getCompanyBasedList": {
				"type": "object",
				"properties": {
					"companyid": {
						"type": "string"
					}

				}
			},
...
			"getCompanyDetails":{
				"type" : "object",
				"properties" : {
					"companyid" : {
						"type" : "string"
					},
					"email" : {
						"type" : "string"
					}
				}
			},
...
			"getCompanyList":{
				"type" : "object",
				"properties" : {

				}
			},
...
			"deleteCompany":{
				 "type": "object",
				 "properties": {
					"companyid": {
						"type": "string"
					}
                }
			},
...
		}
	},
	"paths": {
...
		"/compinfo/addCompany": {
			"post": {
				"tags": [
					"Company Info"
				],
				"summary": "addCompany",
				"requestBody": {
					"description": "addCompany API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/addCompany"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/addCompany"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/addCompany"
						}
					}
				}
			}
		},
...
		"/compinfo/updateCompany": {
			"put": {
				"tags": [
					"Company Info"
				],
				"summary": "updateCompany",
				"requestBody": {
					"description": "updateCompany API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/updateCompany"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/updateCompany"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/updateCompany"
						}
					}
				}
			}
		},
		"/compinfo/deleteCompany": {
			"post": {
				"tags": [
					"Company Info"
				],
				"summary": "deleteCompany",
				"requestBody": {
					"description": "deleteCompany API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/deleteCompany"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/deleteCompany"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/deleteCompany"
						}
					}
				}
			}
		},
...
		"/compinfo/getCompanyBasedList": {
			"post": {
				"tags": [
					"Company Info"
				],
				"summary": "getCompanyBasedList",
				"requestBody": {
					"description": "getCompanyBasedList API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/getCompanyBasedList"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/getCompanyBasedList"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/getCompanyBasedList"
						}
					}
				}
			}
		},
...
	"/compinfo/getCompanyDetails":{
			"post" : {
				"tags": [
					"Company Info"
				],
				"summary": " getCompanyDetails",
				"requestBody": {
					"description": "get getCompanyDetails API",
					"required": true,
					"content": {
						"application/x-www-form-urlencoded": {
							"schema": {
								"$ref": "#/components/schemas/getCompanyDetails"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/getCompanyDetails"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/getCompanyDetails"
						}
					}
				}
			}
		},
		"/compinfo/getCompanyList":{
			"get" : {
				"tags": [
					"Company Info"
				],
				"summary": "Get getCompanyList",
				"parameters": [
					{
						"in": "query",
						"description": "get getCompanyList",

						"schema": {
							"$ref": "#/components/schemas/getCompanyList"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/components/schemas/getCompanyDetails"
						}
					},
					"404": {
						"description": "NOT FOUND",
						"schema": {
							"$ref": "#/components/schemas/getCompanyList"
						}
					}
				}
			}
		}
	}
}